backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-jazzmin==3.0.0
python-dotenv==1.0.1
python-environ==0.4.54
sqlparse==0.5.0
psycopg2==2.9.9
djangorestframework==3.15.2
web3==7.5.0
django-cors-headers==4.6.0
